feat: parsing number with scientific notation
- the builtin parse function parses floats that are represented in
  scientific notation, so why not utilize it?
- this commits adds the ability to enter numbers in scientific notation
  1.1e2   = 1.1 * 10 ^ 2  = 110
  1.1e+2  = 1.1 * 10 ^ 2  = 110
  1.1e-2  = 1.1 * 10 ^ -1 = 0.011

Lets take an example 1.1e+2 here we first parse 1.1 wrt to previous
commits, now we encounter `e`
    1.1e+2        |   1.1e+2
    ^^^ parsed    |      ^ to parse
    Since we now hold a number in hand we parse this `e` as an exponent
    if we didn't have that number at hand then this `e` would have no
    meaning.

Now we have parsed `1.1e` After this there are three possibilities there
can be `+` `-` or any digit
  - If it is a digit then we can parse it as usual
  - If it is `+` or `-` then we have two possibilities, these signs can
    only be after an `e` to signify that this is a sign for the exponent
    otherwise this is simply an addition or subtraction sign.

If we have a number at hand where last we parsed an `e` then this sign
is for the exponent.

If we have a number at hand and we have not lastly parsed `e` then this
is just a normal sign for addition and subtraction. And so we store that
number into stack and move on to do the calculation.